azure-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pyairtable>=3.0.1
Unidecode
python-dotenv
azure-storage-file-datalake 
azure-identity
outscraper>=5.8.0
azure-functions-durable
azure-cosmos>=4.7.0
openai>=1.0.0